MOTE

MOTE is a fast implementation of MOTE RF (Guo et al. 2021), a random forest model that simultaneously estimates individualized treatment effects of multivariate outcomes.

MOTE(
  x.b,
  x.e,
  treat,
  y.b,
  y.e,
  Z,
  num.trees = 300,
  min.node.size = NULL,
  max.depth = NULL,
  num.random.splits = 1,
  minprop = 0.1,
  write.forest = TRUE,
  replace = TRUE,
  sample.fraction = ifelse(replace, 1, 0.632),
  class.weights = NULL,
  keep.inbag = FALSE,
  inbag = NULL,
  holdout = FALSE,
  oob.error = TRUE,
  num.threads = 1,
  verbose = TRUE,
  seed = NULL,
  dependent.variable.name = NULL,
  trt.variable.name = NULL,
  ...
)

Arguments

num.trees
Number of trees.

min.node.size
Minimal node size. Default 1 for classification, 5 for regression, 3 for survival, and 10 for probability.

max.depth
Maximal tree depth. A value of NULL or 0 (the default) corresponds to unlimited depth, 1 to tree stumps (1 split per tree).

num.random.splits
For "extratrees" splitrule.: Number of random splits to consider for each candidate splitting variable.

minprop
For "maxstat" splitrule: Lower quantile of covariate distribution to be considered for splitting.

write.forest
Save MOTE.forest object, required for prediction. Set to FALSE to reduce memory usage if no prediction intended.

replace
Sample with replacement.

sample.fraction
Fraction of observations to sample. Default is 1 for sampling with replacement and 0.632 for sampling without replacement. For classification, this can be a vector of class-specific values.

class.weights
Weights for the outcome classes (in order of the factor levels) in the splitting rule (cost sensitive learning). Classification and probability prediction only. For classification the weights are also applied in the majority vote in terminal nodes.

keep.inbag
Save how often observations are in-bag in each tree.

inbag
Manually set observations per tree. List of size num.trees, containing inbag counts for each observation. Can be used for stratified sampling.

holdout
Hold-out mode. Hold-out all samples with case weight 0 and use these for variable importance and prediction error.

oob.error
Compute OOB prediction error. Set to FALSE to save computation time, e.g. for large survival forests.

num.threads
Number of threads. Default is number of CPUs available.

verbose
Show computation status and estimated runtime.

seed
Random seed. Default is NULL, which generates the seed from R. Set to 0 to ignore the R seed.

dependent.variable.name
Name of dependent variable, needed if no formula given. For survival forests this is the time variable.

trt.variable.name
Name of treatment variable. Treatment variable should be a factor variable

...
Further arguments passed to or from other methods (currently ignored).

formula
Object of class formula or character describing the model to fit. Interaction terms supported only for numerical variables.

data
Training data of class data.frame, matrix, dgCMatrix (Matrix).

save.memory
Use memory saving (but slower) splitting mode. No effect for survival and GWAS data. Warning: This option slows down the tree growing, use only if you encounter memory problems.

x
Predictor data (independent variables), alternative interface to data with formula or dependent.variable.name.

y
Response vector (dependent variable), alternative interface to data with formula or dependent.variable.name. For survival use a Surv() object or a matrix with time and status.

Value

Object of class ranger with elements

forest

Saved forest (If write.forest set to TRUE). Note that the variable IDs in the split.varIDs object do not necessarily represent the column number in R.

predictions

Predicted classes/values, based on out of bag samples (classification and regression only).

variable.importance

Variable importance for each independent variable.

prediction.error

Overall out of bag prediction error. For classification this is the fraction of missclassified samples, for probability estimation the Brier score, for regression the mean squared error and for survival one minus Harrell's C-index.

r.squared

R squared. Also called explained variance or coefficient of determination (regression only). Computed on out of bag data.

confusion.matrix

Contingency table for classes and predictions based on out of bag samples (classification only).

call

Function call.

num.trees

Number of trees.

num.independent.variables

Number of independent variables.

mtry

Value of mtry used.

min.node.size

Value of minimal node size used.

num.samples

Number of samples.

inbag.counts

Number of times the observations are in-bag in the trees.

Details

Forests of extremely randomized trees (Geurts et al. 2006) can be grown.

Note that:

nodes with size smaller than min.node.size can occur, as in original Random Forests.

The mtry variable that randomly selected variables doesn't exist in our framework, as we are using oblique splitting rules.

For factor covriates, a reference coding design matrix will be created accordingly. TODO: it is possible to extend this later by extending to ordered factors.

The importance of variables/features are calculated by accumulating the coefficients in the oblique splitting rules weighted by node sample size.

The model yet to provide handling of missing value. Data containing missing value can promote computation error and warning.

For a large number of variables and data frames as input data the formula interface can be slow or impossible to use. TODO: need improve this part a little bit Alternatively dependent.variable.name and status.variable.name for treatment category or x and y can be used. Use x and y with a matrix for x to avoid conversions and save memory.

TODO: Since our algorithm is oriented to analyze microbiome data, we can directly use phyloseq object. Case weights is not supported temporarily

See https://github.com/boyiguo1/MOTE.RF for the development version.

With recent R versions, multithreading on Windows platforms should just work. If you compile yourself, the new RTools toolchain is required.

References

Guo et al. (Submitted). Estimating Heterogeneous Treatment Effect on Multivariate Responses using Random Forests. https://doi.org/xxxx.

See also

Author

Boyi Guo

Examples

set.seed(1)

p = 10
q = 3

B = create.B(p)
Z = create.Z(p,q)

sim.dat <- sim_MOTE_data(B = B, Z = Z)